+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.4 - unreleased
+
+- Sprite record presets can set a fixed piece size (`subsprite_tiles`), used
+  for SNES OAM tables of 16x16 objects
+
 ## v1.1.3 - 2026-09-22
 
 - Palette files can now be opened as swatches and sliced